+++ src/main/java/ru/yandex/practicum/filmorate/storage/interfaces/FilmStorage.java
@@ -10,5 +10,5 @@ public interface FilmStorage {
     Film update(Film film) throws FilmExistException;
     Film delete(Film film) throws FilmExistException;
     List<Film> getFilms();
-    Film getById(int id) throws FilmExistException;
+    Film getById(long id) throws FilmExistException;
 }
